fixed: 退出逻辑优化

+++ scripts/log-out.js
@@ -1,11 +1,15 @@
 // 退出登录
 (function () {
-  var token = localStorage["auth-token"]
-  if (token) {
-    localStorage.clear()
-    location.reload()
+  chrome.storage.local.clear()
+  // 判断当面页面是否为 "web.okjike.com"
+  if (window.location.host.indexOf('web.okjike.com') > -1) {
+    var token = localStorage["auth-token"]
+    if (token) {
+      localStorage.clear()
+      location.reload()
+    }
+    chrome.runtime.sendMessage({
+      token: null
+    }, null)
   }
-  chrome.runtime.sendMessage({
-    token: null
-  }, null)
 })()
